# Reduced memory allocs hashKernel, countMinSketch.Reset()