## hashKernel

this method gets the upper/lower portions of the hash just using bitwise operations, instead of using a slice.

this reduces the allocs of this method to 0 

before:

```
$ go test -bench=. -benchmem boom.go boom_test.go
goos: darwin
goarch: amd64
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
BenchmarkHashKernel-4           25947562                38.74 ns/op            8 B/op          1 allocs/op
PASS
```

after:

```
$ go test -bench=. -benchmem boom.go boom_test.go
goos: darwin
goarch: amd64
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
BenchmarkHashKernel-4          80436614                12.77 ns/op            0 B/op          0 allocs/op
PASS
```

this should improve performance on methods that use this - e.g. countMinSketch Add/Count

before:

```
BenchmarkCMSAdd-4       19589024                69.54 ns/op            8 B/op          1 allocs/op
BenchmarkCMSCount-4     19352895                67.81 ns/op            8 B/op          1 allocs/op
```

after:

```
BenchmarkCMSAdd-4       41909880                31.24 ns/op            0 B/op          0 allocs/op
BenchmarkCMSCount-4     34261183                37.24 ns/op            0 B/op          0 allocs/op
```

## countMinSketch.Reset()

instead of creating a new matrix (which causes allocs), this re-uses the existing matrix and resets the counters to 0

before:

```
$ go test -bench=BenchmarkCMSReset -benchmem countmin.go countmin_test.go boom.go
goos: darwin
goarch: amd64
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
BenchmarkCMSReset-4         9706            118053 ns/op          663632 B/op          4 allocs/op
PASS
```

after:

```
$ go test -bench=BenchmarkCMSReset -benchmem countmin.go countmin_test.go boom.go
goos: darwin
goarch: amd64
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
BenchmarkCMSReset-4        14294             82448 ns/op               0 B/op          0 allocs/op
PASS
```